pharma_rag_chat_bot/requirement.txt
streamlit
python-dotenv
langchain-core
langchain
langchain-groq
langchain-huggingface
langchain-community
faiss-cpu
sentence-transformers
torch
transformers
PyPDF2
